23.11.28 15:35 commit; 회원탈퇴수정완료

+++ GamersFiles/src/main/java/com/project/controller/UDMService.java
@@ -18,30 +18,43 @@ public class UDMService implements Service {
 		String m_id = request.getParameter("m_id");
 		String g_name = request.getParameter("g_name");
 		GroupDTO group = new GroupDAO().selectGroup(g_name);
-		if (m_id.equals(group.getM_id())) {
-			int groupM = new DeleteDAO().deleteGroupinfo(g_name);
-			if (groupM > 0) {
-				System.out.println("이름 삭제 완료");
-				int deleteG = new GroupDAO().deleteGroup(g_name);
-				if (deleteG > 0) {
-					System.out.println("파티 삭제 완료");
-					int deleteM = new MemberDAO().deleteMember(m_id);
-					if (deleteM > 0) {
-						System.out.println("회원 탈퇴 완료");
-						HttpSession session = request.getSession();
-						session.invalidate();
-						return "GamersMain.jsp";
+		if (group != null) {
+			if (m_id.equals(group.getM_id())) {
+				int groupM = new DeleteDAO().deleteGroupinfo(g_name);
+				if (groupM > 0) {
+					System.out.println("이름 삭제 완료");
+					int deleteG = new GroupDAO().deleteGroup(g_name);
+					if (deleteG > 0) {
+						System.out.println("파티 삭제 완료");
+						int deleteM = new MemberDAO().deleteMember(m_id);
+						if (deleteM > 0) {
+							System.out.println("회원 탈퇴 완료");
+							HttpSession session = request.getSession();
+							session.invalidate();
+							return "GamersMain.jsp";
+						} else {
+							System.out.println("회원 탈퇴 실패");
+							return "GamersMain.jsp";
+						}
 					} else {
-						System.out.println("회원 탈퇴 실패");
+						System.out.println("파티 삭제 실패");
 						return "GamersMain.jsp";
 					}
 				} else {
-					System.out.println("파티 삭제 실패");
+					System.out.println("이름 삭제 실패");
 					return "GamersMain.jsp";
 				}
 			} else {
-				System.out.println("이름 삭제 실패");
-				return "GamersMain.jsp";
+				int cnt = new MemberDAO().deleteMember(m_id);
+				if (cnt > 0) {
+					System.out.println("회원 탈퇴 성공");
+					HttpSession session = request.getSession();
+					session.invalidate();
+					return "GamersMain.jsp";
+				} else {
+					System.out.println("회원 탈퇴 실패");
+					return "GamersMain.jsp";
+				}
 			}
 		} else {
 			int cnt = new MemberDAO().deleteMember(m_id);
